+++ drivers/winmidi/midi_driver_winmidi.cpp
@@ -1,5 +1,5 @@
 /*************************************************************************/
-/*  core_midi.h                                                          */
+/*  midi_driver_winmidi.cpp                                              */
 /*************************************************************************/
 /*                       This file is part of:                           */
 /*                           GODOT ENGINE                                */
@@ -28,35 +28,80 @@
 /* SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                */
 /*************************************************************************/
 
-#ifdef COREMIDI_ENABLED
+#ifdef WINMIDI_ENABLED
 
-#ifndef CORE_MIDI_H
-#define CORE_MIDI_H
+#include "midi_driver_winmidi.h"
 
-#include "core/os/midi_driver.h"
-#include "core/vector.h"
+#include "core/print_string.h"
 
-#include <CoreMIDI/CoreMIDI.h>
-#include <stdio.h>
+void MIDIDriverWinMidi::read(HMIDIIN hMidiIn, UINT wMsg, DWORD_PTR dwInstance, DWORD_PTR dwParam1, DWORD_PTR dwParam2) {
 
-class MIDIDriverCoreMidi : public MIDIDriver {
+	if (wMsg == MIM_DATA) {
+		receive_input_packet((uint64_t)dwParam2, (uint8_t *)&dwParam1, 3);
+	}
+}
 
-	MIDIClientRef client;
-	MIDIPortRef port_in;
+Error MIDIDriverWinMidi::open() {
 
-	Vector<MIDIEndpointRef> connected_sources;
+	for (UINT i = 0; i < midiInGetNumDevs(); i++) {
+		HMIDIIN midi_in;
 
-	static void read(const MIDIPacketList *packet_list, void *read_proc_ref_con, void *src_conn_ref_con);
+		MMRESULT res = midiInOpen(&midi_in, i, (DWORD_PTR)read, (DWORD_PTR)this, CALLBACK_FUNCTION);
+		if (res == MMSYSERR_NOERROR) {
+			midiInStart(midi_in);
+			connected_sources.insert(i, midi_in);
+		} else {
+			char err[256];
+			midiInGetErrorText(res, err, 256);
+			ERR_PRINTS("midiInOpen error: " + String(err));
 
-public:
-	virtual Error open();
-	virtual void close();
+			MIDIINCAPS caps;
+			res = midiInGetDevCaps(i, &caps, sizeof(MIDIINCAPS));
+			if (res == MMSYSERR_NOERROR) {
+				ERR_PRINTS("Can't open MIDI device \"" + String(caps.szPname) + "\", is it being used by another application?");
+			}
+		}
+	}
 
-	PoolStringArray get_connected_inputs();
+	return OK;
+}
 
-	MIDIDriverCoreMidi();
-	virtual ~MIDIDriverCoreMidi();
-};
+PoolStringArray MIDIDriverWinMidi::get_connected_inputs() {
 
-#endif
-#endif
+	PoolStringArray list;
+
+	for (int i = 0; i < connected_sources.size(); i++) {
+		HMIDIIN midi_in = connected_sources[i];
+		UINT id = 0;
+		MMRESULT res = midiInGetID(midi_in, &id);
+		if (res == MMSYSERR_NOERROR) {
+			MIDIINCAPS caps;
+			res = midiInGetDevCaps(i, &caps, sizeof(MIDIINCAPS));
+			if (res == MMSYSERR_NOERROR) {
+				list.push_back(caps.szPname);
+			}
+		}
+	}
+
+	return list;
+}
+
+void MIDIDriverWinMidi::close() {
+
+	for (int i = 0; i < connected_sources.size(); i++) {
+		HMIDIIN midi_in = connected_sources[i];
+		midiInStop(midi_in);
+		midiInClose(midi_in);
+	}
+	connected_sources.clear();
+}
+
+MIDIDriverWinMidi::MIDIDriverWinMidi() {
+}
+
+MIDIDriverWinMidi::~MIDIDriverWinMidi() {
+
+	close();
+}
+
+#endif // WINMIDI_ENABLED
