+++ servers/visual/visual_server_canvas.cpp
@@ -29,7 +29,7 @@
 /*************************************************************************/
 
 #include "visual_server_canvas.h"
-#include "visual_server_global.h"
+#include "visual_server_globals.h"
 #include "visual_server_raster.h"
 #include "visual_server_viewport.h"
