+++ servers/visual/visual_server_globals.cpp
@@ -1,5 +1,5 @@
 /*************************************************************************/
-/*  win_midi.h                                                           */
+/*  visual_server_globals.cpp                                            */
 /*************************************************************************/
 /*                       This file is part of:                           */
 /*                           GODOT ENGINE                                */
@@ -28,34 +28,13 @@
 /* SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                */
 /*************************************************************************/
 
-#ifdef WINMIDI_ENABLED
+#include "visual_server_globals.h"
 
-#ifndef WIN_MIDI_H
-#define WIN_MIDI_H
+RasterizerStorage *VisualServerGlobals::storage = NULL;
+RasterizerCanvas *VisualServerGlobals::canvas_render = NULL;
+RasterizerScene *VisualServerGlobals::scene_render = NULL;
+Rasterizer *VisualServerGlobals::rasterizer = NULL;
 
-#include "core/os/midi_driver.h"
-#include "core/vector.h"
-
-#include <stdio.h>
-#include <windows.h>
-
-#include <mmsystem.h>
-
-class MIDIDriverWinMidi : public MIDIDriver {
-
-	Vector<HMIDIIN> connected_sources;
-
-	static void CALLBACK read(HMIDIIN hMidiIn, UINT wMsg, DWORD_PTR dwInstance, DWORD_PTR dwParam1, DWORD_PTR dwParam2);
-
-public:
-	virtual Error open();
-	virtual void close();
-
-	virtual PoolStringArray get_connected_inputs();
-
-	MIDIDriverWinMidi();
-	virtual ~MIDIDriverWinMidi();
-};
-
-#endif
-#endif
+VisualServerCanvas *VisualServerGlobals::canvas = NULL;
+VisualServerViewport *VisualServerGlobals::viewport = NULL;
+VisualServerScene *VisualServerGlobals::scene = NULL;
