+++ servers/visual/visual_server_viewport.cpp
@@ -32,7 +32,7 @@
 
 #include "core/project_settings.h"
 #include "visual_server_canvas.h"
-#include "visual_server_global.h"
+#include "visual_server_globals.h"
 #include "visual_server_scene.h"
 
 void VisualServerViewport::_draw_viewport(Viewport *p_viewport, ARVRInterface::Eyes p_eye) {
